requirements.txt
blinker==1.9.0
click==8.1.8
colorama==0.4.6
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
MarkupSafe==3.0.2
numpy==2.2.3
pillow==11.1.0
scikit-learn==1.6.1
scipy==1.15.2
threadpoolctl==3.5.0
Werkzeug==3.1.3
